Simplify Milan transport ownership
