+++ drivers/goodix53x5/device/transport.c
@@ -22,8 +22,8 @@
 #include "drivers_api.h"
 #include "driver-private.h"
 #include "device/transport.h"
+#include "device/commands.h"
 #include "device/calibration.h"
-#include "device/scan.h"
 
 #include <string.h>
 
@@ -44,119 +44,53 @@
 #define GOODIX_PROTO_CMD_BYTE(category, command) \
   (((category) << 4) | ((command) << 1))
 
-/* Command sub-SSM */
 typedef enum {
-  GOODIX_CMD_SEND = 0,
-  GOODIX_CMD_RECV_ACK,
-  GOODIX_CMD_VALIDATE_ACK,
-  GOODIX_CMD_RECV_DATA,
-  GOODIX_CMD_NUM_STATES,
-} GoodixCmdState;
+  GOODIX_TRANSPORT_IDLE,
+  GOODIX_TRANSPORT_JOIN_IDLE,
+  GOODIX_TRANSPORT_SEND,
+  GOODIX_TRANSPORT_ACK,
+  GOODIX_TRANSPORT_RESPONSE,
+  GOODIX_TRANSPORT_EVENT,
+  GOODIX_TRANSPORT_REPLY,
+  GOODIX_TRANSPORT_STOPPING,
+} GoodixTransportPhase;
 
-typedef struct
+/* One owner survives every submitted callback. JOIN_IDLE reserves a command,
+ * but the winning old callback is still processed under idle policy. */
+struct _GoodixTransport
 {
-  guint64                     token;
-  guint                       timeout_ms;
-  gint64                      deadline_us;
-  gboolean                    fixed_deadline;
-  GCancellable               *cancellable;
-  GoodixRecvCancelledCallback cancelled_cb;
-  gpointer                    cancelled_data;
-} GoodixRecvOperation;
-
-typedef struct
-{
+  FpDevice                 *dev;
+  GoodixTransportPhase      phase;
+  GoodixTransportDone       done;
+  gpointer                  data;
+  GoodixTransportJoined     joined;
+  gpointer                  joined_data;
+  GCancellable             *cancellable;
+  guint                     timeout_ms;
+  gint64                    deadline_us;
+  GoodixProfile9FdtWaitMode event_mode;
   GoodixCmd                 cmd;
+  gboolean                  expect_data;
   gboolean                  idle_after_ack;
-  FpiSsm                   *parent_ssm;
   GoodixProfile9FdtWaitMode cancelled_fdt_mode;
   gboolean                  retry_mode;
   guint8                    response_bit;
   guint8                    ack_status;
-  GoodixCmdResultCallback   result_cb;
   guint                     attempt;
-  GoodixCmdState            phase;
-  gint64                    ack_deadline_us;
   guint                     ack_timeout_ms;
   guint                     response_timeout_ms;
-} GoodixCmdOperation;
+};
 
-typedef struct
+static void goodix_transport_send (GoodixTransport *operation);
+static void goodix_transport_receive (GoodixTransport *operation, guint timeout);
+static void goodix_transport_complete (GoodixTransport *operation, GError *error);
+
+static gboolean
+goodix_transport_is_idle (GoodixTransport *operation)
 {
-  FpDevice                *dev;
-  GCancellable            *cancel;
-  GoodixIdleJoinedCallback joined;
-  gpointer                 joined_data;
-} GoodixIdleRecv;
-
-static void
-goodix_idle_recv_free (GoodixIdleRecv *idle)
-{
-  g_object_unref (idle->cancel);
-  g_object_unref (idle->dev);
-  g_free (idle);
-}
-
-static void
-goodix_idle_recv_done (FpiSsm *ssm, FpDevice *dev, GError *error)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixIdleRecv *idle = fpi_ssm_get_data (ssm);
-
-  self->idle_rx_ssm = NULL;
-  self->rx_idle_partial = self->rx.len && !goodix_proto_rx_complete (&self->rx);
-  if (error && !g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED))
-    {
-      fp_dbg ("Idle receive stopped: %s", error->message);
-      self->needs_reinit = TRUE;
-    }
-  g_clear_error (&error);
-  fp_dbg ("Idle receive joined; partial=%d", self->rx_idle_partial);
-  if (idle->joined)
-    idle->joined (dev, idle->joined_data);
-}
-
-static void
-goodix_idle_recv_handler (FpiSsm *ssm, FpDevice *dev)
-{
-  GoodixIdleRecv *idle = fpi_ssm_get_data (ssm);
-
-  goodix_recv_start (ssm, dev, 0, idle->cancel);
-}
-
-static void
-goodix_idle_recv_start (FpDevice *dev)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixIdleRecv *idle = g_new0 (GoodixIdleRecv, 1);
-
-  g_assert (!self->rx_active && !self->idle_rx_ssm);
-  idle->dev = g_object_ref (dev);
-  idle->cancel = g_cancellable_new ();
-  self->idle_rx_ssm = fpi_ssm_new (dev, goodix_idle_recv_handler, 1);
-  fpi_ssm_set_data (self->idle_rx_ssm, idle, (GDestroyNotify) goodix_idle_recv_free);
-  fp_dbg ("Idle receive started after EC-off ACK");
-  fpi_ssm_start (self->idle_rx_ssm, goodix_idle_recv_done);
-}
-
-void
-goodix_idle_recv_stop (FpDevice *dev, GoodixIdleJoinedCallback joined,
-                       gpointer data)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-
-  if (self->idle_rx_ssm)
-    {
-      GoodixIdleRecv *idle = fpi_ssm_get_data (self->idle_rx_ssm);
-      g_assert (!idle->joined);
-      idle->joined = joined;
-      idle->joined_data = data;
-      g_cancellable_cancel (idle->cancel);
-    }
-  else
-    {
-      joined (dev, data);
-    }
+  return operation->phase == GOODIX_TRANSPORT_IDLE ||
+         operation->phase == GOODIX_TRANSPORT_JOIN_IDLE ||
+         operation->phase == GOODIX_TRANSPORT_STOPPING;
 }
 
 /* Native budgets count Sleep(1) ACK polls and 50-ms response waits. GUsb uses
@@ -164,7 +98,7 @@ goodix_idle_recv_stop (FpDevice *dev, GoodixIdleJoinedCallback joined,
  * complete packet reception, beginning after write/ACK completion respectively.
  * Other commands retain their unaudited existing timing policy. */
 static void
-goodix_cmd_set_budgets (GoodixCmdOperation *operation)
+goodix_cmd_set_budgets (GoodixTransport *operation)
 {
   guint8 command = GOODIX_PROTO_CMD_BYTE (operation->cmd.category,
                                           operation->cmd.command);
@@ -233,24 +167,14 @@ goodix_mode_ack_bit (guint8 cmd_byte)
 }
 
 static const char *
-goodix_cmd_phase_name (GoodixCmdState phase)
+goodix_cmd_phase_name (GoodixTransportPhase phase)
 {
-  switch (phase)
-    {
-    case GOODIX_CMD_SEND:
-      return "send";
-
-    case GOODIX_CMD_RECV_ACK:
-    case GOODIX_CMD_VALIDATE_ACK:
-      return "ACK";
-
-    case GOODIX_CMD_RECV_DATA:
-      return "data";
-
-    case GOODIX_CMD_NUM_STATES:
-      return "unknown";
-    }
-  g_assert_not_reached ();
+  if (phase == GOODIX_TRANSPORT_SEND)
+    return "send";
+  if (phase == GOODIX_TRANSPORT_ACK)
+    return "ACK";
+  g_assert (phase == GOODIX_TRANSPORT_RESPONSE);
+  return "data";
 }
 
 /* Only a transport timeout or send I/O failure in the first transaction is
@@ -258,11 +182,11 @@ goodix_cmd_phase_name (GoodixCmdState phase)
  * keep their existing failure policy. The parent and command owner never
  * change between attempts. */
 static gboolean
-goodix_cmd_native_zero (GoodixCmdOperation *operation, const GError *error)
+goodix_cmd_native_zero (GoodixTransportPhase phase, const GError *error)
 {
   if (g_error_matches (error, G_USB_DEVICE_ERROR, G_USB_DEVICE_ERROR_TIMED_OUT))
     return TRUE;
-  if (operation->phase != GOODIX_CMD_SEND || !error || error->domain != G_USB_DEVICE_ERROR)
+  if (phase != GOODIX_TRANSPORT_SEND || !error || error->domain != G_USB_DEVICE_ERROR)
     return FALSE;
   /* Native maps negative writes to zero. These GUsb codes include transfer
    * error, endpoint stall and ordinary backend failure. Host preconditions,
@@ -274,33 +198,24 @@ goodix_cmd_native_zero (GoodixCmdOperation *operation, const GError *error)
 }
 
 static gboolean
-goodix_cmd_retry (FpDevice *dev,
-                  FpiSsm   *ssm,
-                  GError   *error)
+goodix_cmd_retry (GoodixTransport *operation, GError *error)
 {
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixCmdOperation *operation;
+  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (operation->dev);
+  GoodixTransportPhase phase = operation->phase;
 
-  if (self->cmd_ssm != ssm || self->rx_active)
-    return FALSE;
-  operation = fpi_ssm_get_data (ssm);
-  if (self->cmd_owner != operation->parent_ssm ||
-      !operation->retry_mode || operation->attempt != 1 ||
-      (operation->phase != GOODIX_CMD_SEND &&
-       operation->phase != GOODIX_CMD_RECV_ACK &&
-       !(operation->response_bit && operation->phase == GOODIX_CMD_RECV_DATA)) ||
-      !goodix_cmd_native_zero (operation, error))
+  if (!operation->retry_mode || operation->attempt != 1 ||
+      (phase != GOODIX_TRANSPORT_SEND && phase != GOODIX_TRANSPORT_ACK &&
+       !(operation->response_bit && phase == GOODIX_TRANSPORT_RESPONSE)) ||
+      !goodix_cmd_native_zero (phase, error))
     return FALSE;
 
-  fp_dbg ("Retrying command cat=0x%02x cmd=0x%02x phase=%s attempt=1 parent-state=%d: %s",
+  fp_dbg ("Retrying command cat=0x%02x cmd=0x%02x phase=%s attempt=1: %s",
           operation->cmd.category, operation->cmd.command,
-          goodix_cmd_phase_name (operation->phase),
-          fpi_ssm_get_cur_state (operation->parent_ssm), error->message);
+           goodix_cmd_phase_name (phase), error->message);
   self->retried_mode_acks |= goodix_mode_ack_bit (
     GOODIX_PROTO_CMD_BYTE (operation->cmd.category, operation->cmd.command));
-  operation->ack_deadline_us = 0;
   g_error_free (error);
-  fpi_ssm_jump_to_state (ssm, GOODIX_CMD_SEND);
+  goodix_transport_send (operation);
   return TRUE;
 }
 
@@ -313,45 +228,52 @@ goodix_deadline_remaining (gint64 deadline_us)
   return remaining > 0 ? (guint) ((remaining + 999) / 1000) : 0;
 }
 
-static void
-goodix_mark_coordinator_io_failure (FpiDeviceGoodix53x5 *self,
-                                    const GError         *error)
-{
-  if (self->cmd_ssm && ((GoodixCmdOperation *) fpi_ssm_get_data (self->cmd_ssm))->result_cb)
-    return; /* The composite command owns its intermediate result. */
-  if (self->profile9_fdt.owner &&
-      !g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED))
-    self->needs_reinit = TRUE;
-}
-
-static void
-goodix_cmd_operation_free (GoodixCmdOperation *operation)
-{
-  if (!operation)
-    return;
-
-  g_free (operation->cmd.payload);
-  g_free (operation);
-}
-
 static gboolean
-goodix_validate_ack_for_cmd (FpDevice        *dev,
-                             const GoodixCmd *cmd,
+goodix_validate_ack_for_cmd (FpDevice       *dev,
+                             GoodixTransport *operation,
                              guint8          *status,
                              GError         **error)
 {
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
   guint8 category, command;
   const guint8 *payload;
   gsize payload_len;
-  guint8 expected_cmd_byte = GOODIX_PROTO_CMD_BYTE (cmd->category, cmd->command);
+  guint8 expected_cmd_byte = GOODIX_PROTO_CMD_BYTE (operation->cmd.category,
+                                                   operation->cmd.command);
 
-  if (!goodix_proto_rx_parse (&self->rx, &category, &command,
-                              &payload, &payload_len))
+  if (!goodix_parse_reply (dev, &category, &command,
+                           &payload, &payload_len, NULL))
     {
       g_set_error_literal (error, FP_DEVICE_ERROR, FP_DEVICE_ERROR_PROTO,
                            "Failed to parse ACK");
       return FALSE;
+    }
+
+  if (category == GOODIX_PROTO_CATEGORY_FDT &&
+      operation->cancelled_fdt_mode != GOODIX_PROFILE9_FDT_WAIT_NONE)
+    {
+      GoodixFdtEventType type;
+      GoodixProfile9FdtEvent event;
+      guint8 expected_command = operation->cancelled_fdt_mode == GOODIX_PROFILE9_FDT_WAIT_DOWN
+                                  ? GOODIX_PROTO_CMD_FDT_DOWN : GOODIX_PROTO_CMD_FDT_UP;
+
+      if (command != expected_command || payload_len != GOODIX_FDT_EVENT_PAYLOAD_LEN)
+        {
+          g_set_error (error, FP_DEVICE_ERROR, FP_DEVICE_ERROR_PROTO,
+                       "Unexpected FDT packet during cleanup: armed=0x%02x cat=0x%02x cmd=0x%02x len=%zu",
+                       expected_command, category, command, payload_len);
+          return FALSE;
+        }
+      if (!goodix_cmd_parse_fdt_event (dev, operation->cancelled_fdt_mode,
+                                      &type, &event, error))
+        return FALSE;
+      /* Stopping invalidates notification, not parser-side base publication.
+       * A zero status continues ACK reception with the original deadline. */
+      goodix_recv_apply_fdt_event (dev, type, &event);
+      operation->cancelled_fdt_mode = GOODIX_PROFILE9_FDT_WAIT_NONE;
+      fp_dbg ("Drained cancelled FDT event before sleep ACK: mode=0x%02x irq=0x%04x",
+              expected_command, event.irq);
+      *status = 0;
+      return TRUE;
     }
 
   if (category != GOODIX_PROTO_CATEGORY_ACK ||
@@ -376,50 +298,6 @@ goodix_validate_ack_for_cmd (FpDevice        *dev,
   return TRUE;
 }
 
-static gboolean
-goodix_try_drain_cancelled_fdt (FpDevice           *dev,
-                                GoodixCmdOperation *operation,
-                                GError            **error)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  guint8 category, command;
-  guint8 expected_command;
-  const guint8 *payload;
-  gsize payload_len;
-  guint16 irq;
-  GoodixFdtEventType type;
-  GoodixProfile9FdtEvent event;
-
-  if (!goodix_proto_rx_parse (&self->rx, &category, &command,
-                              &payload, &payload_len))
-    return FALSE;
-  if (category != GOODIX_PROTO_CATEGORY_FDT)
-    return FALSE;
-
-  expected_command =
-    operation->cancelled_fdt_mode == GOODIX_PROFILE9_FDT_WAIT_DOWN
-      ? GOODIX_PROTO_CMD_FDT_DOWN : GOODIX_PROTO_CMD_FDT_UP;
-  if (command != expected_command || payload_len != GOODIX_FDT_EVENT_PAYLOAD_LEN)
-    {
-      g_set_error (error, FP_DEVICE_ERROR, FP_DEVICE_ERROR_PROTO,
-                   "Unexpected FDT packet during cleanup: armed=0x%02x cat=0x%02x cmd=0x%02x len=%zu",
-                   expected_command, category, command, payload_len);
-      return FALSE;
-    }
-
-  if (!goodix_cmd_parse_fdt_event (dev, operation->cancelled_fdt_mode,
-                                  &type, &event, error))
-    return FALSE;
-  /* Stopping the worker invalidates notification, not parser-side base
-   * publication. Keep the event local so cleanup cannot restart dispatch. */
-  goodix_recv_apply_fdt_event (dev, type, &event);
-  irq = event.irq;
-  operation->cancelled_fdt_mode = GOODIX_PROFILE9_FDT_WAIT_NONE;
-  fp_dbg ("Drained cancelled FDT event before sleep ACK: mode=0x%02x irq=0x%04x",
-          expected_command, irq);
-  return TRUE;
-}
-
 /* ========================================================================
  * USB I/O helpers
  * ======================================================================== */
@@ -430,103 +308,77 @@ goodix_tx_cb (FpiUsbTransfer *transfer,
               gpointer        user_data,
               GError         *error)
 {
+  GoodixTransport *operation = user_data;
+
   if (error)
     {
-      if (goodix_cmd_retry (dev, transfer->ssm, error))
-        return;
-      /* Cancellation can interrupt a partially sent command. Cleanup using
-       * that same action token may also be prevented from establishing sleep. */
-      if (g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED) ||
-          g_error_matches (error, G_USB_DEVICE_ERROR, G_USB_DEVICE_ERROR_CANCELLED))
-        FPI_DEVICE_GOODIX53X5 (dev)->needs_reinit = TRUE;
-      goodix_mark_coordinator_io_failure (FPI_DEVICE_GOODIX53X5 (dev), error);
-      fpi_ssm_mark_failed (transfer->ssm, error);
+      goodix_transport_complete (operation, error);
       return;
     }
 
-  fpi_ssm_next_state (transfer->ssm);
+  operation->phase = GOODIX_TRANSPORT_ACK;
+  operation->deadline_us = g_get_monotonic_time () + operation->ack_timeout_ms * 1000LL;
+  if ((operation->cmd.category == 0 && operation->cmd.command == 0) ||
+      (operation->cmd.category == 0x0a && operation->cmd.command == 4))
+    operation->cancellable = g_object_ref (fpi_device_get_cancellable (dev));
+  goodix_transport_receive (operation, operation->ack_timeout_ms);
 }
 
-/**
- * Send a complete protocol message, splitting into USB chunks.
- * Advances the SSM on completion.
- */
+/* Submit OUT only after the previous IN callback has joined. The ACK clock
+ * starts in goodix_tx_cb, never while the write is pending. */
 static void
-goodix_send_message (FpiSsm   *ssm,
-                     FpDevice *dev,
-                     guint8    category,
-                     guint8    command,
-                     const guint8 *payload,
-                     gsize     payload_len,
-                     gboolean  use_checksum)
+goodix_transport_send (GoodixTransport *operation)
 {
+  FpDevice *dev = operation->dev;
+  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
+  GoodixCmd *cmd = &operation->cmd;
   gsize msg_len;
   guint8 *msg;
   FpiUsbTransfer *transfer;
   guint8 cmd_byte;
 
-  msg = goodix_proto_build_message (category, command, payload, payload_len,
-                                    use_checksum, &msg_len);
+  operation->phase = GOODIX_TRANSPORT_SEND;
+  operation->attempt++;
+  g_clear_object (&operation->cancellable);
+  self->command_response_ready &= ~operation->response_bit;
+  if (operation->response_bit)
+    self->retried_mode_acks |= goodix_mode_ack_bit (
+      GOODIX_PROTO_CMD_BYTE (cmd->category, cmd->command));
+  msg = goodix_proto_build_message (cmd->category, cmd->command,
+                                    cmd->payload, cmd->payload_len,
+                                    cmd->use_checksum, &msg_len);
   cmd_byte = msg[0];
 
   /* The transport uses 64-byte USB writes. Continuation chunks prepend
    * cmd_byte | 1 and carry up to 63 more bytes of message data. */
 
-  gsize total_chunks = 0;
-  gsize padded_len = 0;
-
-  /* Calculate how many chunks we need */
-  if (msg_len <= GOODIX_USB_CHUNK_SIZE)
-    {
-      total_chunks = 1;
-      padded_len = GOODIX_USB_CHUNK_SIZE;
-    }
-  else
-    {
-      /* First chunk: 64 bytes of message data */
-      gsize remaining = msg_len - GOODIX_USB_CHUNK_SIZE;
-      /* Each continuation chunk carries 63 bytes of data (1 byte for marker) */
-      gsize cont_chunks = (remaining + 62) / 63;
-      total_chunks = 1 + cont_chunks;
-      padded_len = total_chunks * GOODIX_USB_CHUNK_SIZE;
-    }
-
+  gsize first_len = MIN (msg_len, GOODIX_USB_CHUNK_SIZE);
+  gsize total_chunks = 1 + (msg_len - first_len + 62) / 63;
+  gsize padded_len = total_chunks * GOODIX_USB_CHUNK_SIZE;
   guint8 *chunked = g_malloc0 (padded_len);
 
-  if (total_chunks == 1)
+  memcpy (chunked, msg, first_len);
+  gsize src_offset = first_len;
+  for (gsize dst_offset = GOODIX_USB_CHUNK_SIZE;
+       dst_offset < padded_len; dst_offset += GOODIX_USB_CHUNK_SIZE)
     {
-      memcpy (chunked, msg, msg_len);
-    }
-  else
-    {
-      /* First chunk */
-      memcpy (chunked, msg, GOODIX_USB_CHUNK_SIZE);
+      gsize data_in_chunk = MIN (63, msg_len - src_offset);
 
-      gsize src_offset = GOODIX_USB_CHUNK_SIZE;
-      gsize dst_offset = GOODIX_USB_CHUNK_SIZE;
-
-      for (gsize chunk = 1; chunk < total_chunks; chunk++)
-        {
-          chunked[dst_offset] = cmd_byte | 1;
-          gsize data_in_chunk = MIN (63, msg_len - src_offset);
-          if (data_in_chunk > 0)
-            memcpy (chunked + dst_offset + 1, msg + src_offset, data_in_chunk);
-          src_offset += data_in_chunk;
-          dst_offset += GOODIX_USB_CHUNK_SIZE;
-        }
+      chunked[dst_offset] = cmd_byte | 1;
+      memcpy (chunked + dst_offset + 1, msg + src_offset, data_in_chunk);
+      src_offset += data_in_chunk;
     }
 
   g_free (msg);
 
   transfer = fpi_usb_transfer_new (dev);
-  transfer->ssm = ssm;
   fpi_usb_transfer_fill_bulk_full (transfer, GOODIX_EP_OUT,
                                    chunked, padded_len, g_free);
   /* Native writes have no timeout. The active libfprint action owns device
    * lifetime and cancellation until this callback joins; Linux USB removal
    * independently completes in-flight transfers with NO_DEVICE. */
   fpi_usb_transfer_submit (transfer, 0, fpi_device_get_cancellable (dev),
-                           goodix_tx_cb, NULL);
+                           goodix_tx_cb, operation);
 }
 
 /* Forward declarations */
@@ -536,93 +388,103 @@ static void goodix_rx_cb (FpiUsbTransfer *transfer,
                           GError         *error);
 
 static void
-goodix_recv_operation_free (GoodixRecvOperation *operation)
+goodix_transport_free (GoodixTransport *operation)
 {
   g_clear_object (&operation->cancellable);
+  g_free (operation->cmd.payload);
+  g_object_unref (operation->dev);
   g_free (operation);
 }
 
-static gboolean
-goodix_recv_operation_finish (FpDevice           *dev,
-                               FpiSsm             *ssm,
-                               GoodixRecvOperation *operation)
+static GoodixTransport *
+goodix_transport_new (FpDevice             *dev,
+                      GoodixTransportPhase  phase,
+                      GoodixTransportDone   done,
+                      gpointer              data)
 {
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
+  GoodixTransport *operation = g_new0 (GoodixTransport, 1);
 
-  if (!self->rx_active || self->rx_owner != ssm ||
-      self->rx_token != operation->token)
-    return FALSE;
-
-  self->rx_active = FALSE;
-  self->rx_owner = NULL;
-  return TRUE;
+  g_assert (!self->transport);
+  operation->dev = g_object_ref (dev);
+  operation->phase = phase;
+  operation->done = done;
+  operation->data = data;
+  self->transport = operation;
+  return operation;
 }
 
-static gboolean
-goodix_recv_start_full (FpiSsm                     *ssm,
-                        FpDevice                   *dev,
-                        guint                       timeout_ms,
-                        GCancellable               *cancellable,
-                        GoodixRecvCancelledCallback cancelled_cb,
-                        gpointer                    cancelled_data)
+/* Start a new packet wait, retaining only a partial inherited from idle. */
+static void
+goodix_transport_receive (GoodixTransport *operation, guint timeout)
 {
+  FpDevice *dev = operation->dev;
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixRecvOperation *operation;
   FpiUsbTransfer *transfer;
 
-  if (self->rx_active ||
-      (self->cmd_owner != NULL && self->cmd_ssm != ssm))
-    return FALSE;
-
-  operation = g_new0 (GoodixRecvOperation, 1);
-  operation->token = ++self->rx_token;
-  if (operation->token == 0)
-    operation->token = ++self->rx_token;
-  operation->timeout_ms = timeout_ms;
-  operation->deadline_us = timeout_ms ?
-                           g_get_monotonic_time () + timeout_ms * 1000LL : 0;
-  if (self->cmd_ssm == ssm)
-    {
-      GoodixCmdOperation *cmd_operation = fpi_ssm_get_data (ssm);
-
-      if (cmd_operation->phase == GOODIX_CMD_RECV_ACK)
-        {
-          operation->fixed_deadline = TRUE;
-          operation->deadline_us = cmd_operation->ack_deadline_us;
-        }
-      else if (cmd_operation->phase == GOODIX_CMD_RECV_DATA &&
-                cmd_operation->response_bit)
-        {
-          operation->fixed_deadline = TRUE;
-        }
-    }
-  operation->cancellable = cancellable ? g_object_ref (cancellable) : NULL;
-  operation->cancelled_cb = cancelled_cb;
-  operation->cancelled_data = cancelled_data;
-
+  operation->timeout_ms = timeout;
   if (!self->rx_idle_partial)
     goodix_proto_rx_reset (&self->rx);
-  self->rx_active = TRUE;
-  self->rx_owner = ssm;
+  self->reply_valid = FALSE;
 
   transfer = fpi_usb_transfer_new (dev);
-  transfer->ssm = ssm;
   fpi_usb_transfer_fill_bulk (transfer, GOODIX_EP_IN, GOODIX_USB_CHUNK_SIZE);
-  fpi_usb_transfer_submit (transfer, timeout_ms, cancellable,
+  fpi_usb_transfer_submit (transfer, timeout, operation->cancellable,
                            goodix_rx_cb, operation);
-  return TRUE;
 }
 
-void
-goodix_recv_start (FpiSsm       *ssm,
-                   FpDevice     *dev,
-                   guint         timeout_ms,
-                   GCancellable *cancellable)
+static void
+goodix_transport_complete (GoodixTransport *operation, GError *error)
 {
-  if (!goodix_recv_start_full (ssm, dev, timeout_ms, cancellable, NULL, NULL))
-    fpi_ssm_mark_failed (
-      ssm, fpi_device_error_new_msg (FP_DEVICE_ERROR_BUSY,
-                                     "A device receive is already active"));
+  g_autoptr(FpDevice) dev = g_object_ref (operation->dev);
+  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
+  GoodixTransportDone done = operation->done;
+  gpointer data = operation->data;
+  GoodixTransportJoined joined = operation->joined;
+  gpointer joined_data = operation->joined_data;
+  gboolean idle_after_ack = !error && operation->idle_after_ack;
+  GoodixTransportResult result = {
+    .ack_status = operation->ack_status,
+    .ordinary_exhaustion = error && goodix_cmd_native_zero (operation->phase, error),
+    .write_cancelled = operation->phase == GOODIX_TRANSPORT_SEND &&
+      (g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED) ||
+       g_error_matches (error, G_USB_DEVICE_ERROR, G_USB_DEVICE_ERROR_CANCELLED)),
+  };
+
+  if (goodix_transport_is_idle (operation))
+    {
+      self->rx_idle_partial = self->rx.len && !goodix_proto_rx_complete (&self->rx);
+      if (error && !g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED))
+        {
+          fp_dbg ("Idle receive stopped: %s", error->message);
+          self->needs_reinit = TRUE;
+        }
+      g_clear_error (&error);
+      if (operation->phase == GOODIX_TRANSPORT_JOIN_IDLE)
+        {
+          goodix_transport_send (operation);
+          return;
+        }
+    }
+  else if (error && goodix_cmd_retry (operation, error))
+    return;
+  else if (error && operation->attempt)
+    g_prefix_error (&error, "Command cat=0x%02x cmd=0x%02x phase=%s attempt=%u: ",
+                    operation->cmd.category, operation->cmd.command,
+                    goodix_cmd_phase_name (operation->phase), operation->attempt);
+
+  self->transport = NULL;
+  goodix_transport_free (operation);
+  if (idle_after_ack)
+    {
+      GoodixTransport *idle = goodix_transport_new (dev, GOODIX_TRANSPORT_IDLE, NULL, NULL);
+      idle->cancellable = g_cancellable_new ();
+      goodix_transport_receive (idle, 0);
+    }
+  if (done)
+    done (dev, &result, error, data);
+  if (joined)
+    joined (dev, joined_data);
 }
 
 static void
@@ -632,39 +494,16 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
               GError         *error)
 {
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixRecvOperation *operation = user_data;
+  GoodixTransport *operation = user_data;
   FpiUsbTransfer *next;
+  gboolean fixed_deadline = operation->phase == GOODIX_TRANSPORT_ACK ||
+    (operation->phase == GOODIX_TRANSPORT_RESPONSE && operation->response_bit);
 
-  if (!self->rx_active || self->rx_owner != transfer->ssm ||
-      self->rx_token != operation->token)
-    {
-      fp_err ("Discarding callback for an inactive device receive");
-      g_clear_error (&error);
-      goodix_recv_operation_free (operation);
-      return;
-    }
+  g_assert (self->transport == operation);
 
   if (error)
     {
-      goodix_recv_operation_finish (dev, transfer->ssm, operation);
-
-      if (g_error_matches (error, G_IO_ERROR, G_IO_ERROR_CANCELLED) &&
-          operation->cancelled_cb)
-        {
-          GoodixRecvCancelledCallback cancelled_cb = operation->cancelled_cb;
-          gpointer cancelled_data = operation->cancelled_data;
-
-          goodix_recv_operation_free (operation);
-          cancelled_cb (transfer->ssm, dev, error, cancelled_data);
-          return;
-        }
-
-      goodix_recv_operation_free (operation);
-      if (!goodix_cmd_retry (dev, transfer->ssm, error))
-        {
-          goodix_mark_coordinator_io_failure (self, error);
-          fpi_ssm_mark_failed (transfer->ssm, error);
-        }
+      goodix_transport_complete (operation, error);
       return;
     }
 
@@ -672,7 +511,7 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
    * zero-length-read timeout policy. */
   if (transfer->actual_length == 0)
     {
-      if (!operation->fixed_deadline)
+      if (!fixed_deadline)
         operation->deadline_us = operation->timeout_ms ?
                                  g_get_monotonic_time () + operation->timeout_ms * 1000LL : 0;
       goto receive_more;
@@ -681,33 +520,38 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
   if (!goodix_proto_rx_feed_chunk (&self->rx, transfer->buffer,
                                    transfer->actual_length))
     {
-      goodix_recv_operation_finish (dev, transfer->ssm, operation);
-      goodix_recv_operation_free (operation);
-      fpi_ssm_mark_failed (transfer->ssm,
-                           fpi_device_error_new_msg (FP_DEVICE_ERROR_PROTO,
-                                                     "Protocol reassembly error"));
-      return;
+      error = fpi_device_error_new_msg (FP_DEVICE_ERROR_PROTO,
+                                        "Protocol reassembly error");
+      goto protocol_error;
     }
 
   if (goodix_proto_rx_complete (&self->rx))
     {
-      guint8 category, command;
-      const guint8 *payload;
-      gsize payload_len;
-      gboolean expected_ack = FALSE;
-      gboolean idle_packet = self->idle_rx_ssm == transfer->ssm || self->rx_idle_partial;
+      gboolean idle_packet = goodix_transport_is_idle (operation) || self->rx_idle_partial;
+      GoodixTransport *current = !idle_packet &&
+        (operation->phase == GOODIX_TRANSPORT_ACK || operation->phase == GOODIX_TRANSPORT_RESPONSE)
+        ? operation : NULL;
+      gboolean ack_wait = current && operation->phase == GOODIX_TRANSPORT_ACK;
 
-      if (goodix_proto_rx_parse (&self->rx, &category, &command,
-                                 &payload, &payload_len))
+      self->reply_valid = goodix_proto_rx_parse (
+        &self->rx, &self->reply_category, &self->reply_command,
+        &self->reply_payload, &self->reply_payload_len);
+      guint8 category = self->reply_category, command = self->reply_command;
+      const guint8 *payload = self->reply_payload;
+      gsize payload_len = self->reply_payload_len;
+      gboolean ack_packet = self->reply_valid && category == GOODIX_PROTO_CATEGORY_ACK &&
+                            command == GOODIX_PROTO_CMD_ACK && payload_len >= 2;
+      gboolean expected_ack = ack_packet && ack_wait && payload[0] ==
+                              GOODIX_PROTO_CMD_BYTE (current->cmd.category, current->cmd.command);
+
+      if (self->reply_valid)
         {
           guint8 bit = category == 3 && command == 3 ? 1 :
                        category == 9 && command == 0 ? 2 :
                        category == 0x0a && (command == 0 || command == 1 || command == 4) ? 4 : 0;
-          GoodixCmdOperation *current = !idle_packet && self->cmd_ssm == transfer->ssm ?
-                                        fpi_ssm_get_data (transfer->ssm) : NULL;
           gboolean shared = bit == 4 || (category == 0x0a && command == 3) ||
                             category == 8 || category == 0x0e || category == 0x0f;
-          gboolean current_data = current && current->phase == GOODIX_CMD_RECV_DATA &&
+          gboolean current_data = current && operation->phase == GOODIX_TRANSPORT_RESPONSE &&
                                   current->cmd.category == category && current->cmd.command == command;
 
           /* DataFromDevice publishes these shared stores without a matching
@@ -740,18 +584,16 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
                 {
                   if (payload_len < sizeof (self->manual_response))
                     {
-                      goodix_recv_operation_finish (dev, transfer->ssm, operation);
-                      goodix_recv_operation_free (operation);
-                      fpi_ssm_mark_failed (transfer->ssm, fpi_device_error_new_msg (
-                        FP_DEVICE_ERROR_PROTO, "Manual FDT reply is too short"));
-                      return;
+                      error = fpi_device_error_new_msg (
+                        FP_DEVICE_ERROR_PROTO, "Manual FDT reply is too short");
+                      goto protocol_error;
                     }
                   memcpy (self->manual_response, payload, sizeof (self->manual_response));
                 }
               /* Configuration publishes only an event, never a success byte.
                * Response slots are independent of ACK reception. */
               self->command_response_ready |= bit;
-              if (!current || current->phase != GOODIX_CMD_RECV_DATA ||
+              if (!current || operation->phase != GOODIX_TRANSPORT_RESPONSE ||
                   (current->response_bit != bit && !current_data))
                 {
                   goodix_proto_rx_reset (&self->rx);
@@ -773,19 +615,18 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
                                                         "Unexpected FDT event length");
               if (event_error || !goodix_cmd_parse_fdt_event (dev, mode, &type, &event, &event_error))
                 {
-                  goodix_recv_operation_finish (dev, transfer->ssm, operation);
-                  goodix_recv_operation_free (operation);
-                  fpi_ssm_mark_failed (transfer->ssm, g_steal_pointer (&event_error));
-                  return;
+                  error = g_steal_pointer (&event_error);
+                  goto protocol_error;
                 }
               /* Every native parser mutation precedes replacement of the
                * latest worker notification, including during config/manual. */
               goodix_recv_apply_fdt_event (dev, type, &event);
               if (!idle_packet)
                 {
-                  self->pending_fdt_packet_len = self->rx.expected;
-                  self->pending_fdt_mode = mode;
-                  memcpy (self->pending_fdt_packet, self->rx.buf, self->rx.expected);
+                  self->pending_fdt.event = event;
+                  self->pending_fdt.type = type;
+                  memcpy (self->pending_fdt.prior_down, self->fdt_prior_down,
+                          sizeof (self->pending_fdt.prior_down));
                 }
               goodix_proto_rx_reset (&self->rx);
               goto receive_more;
@@ -808,64 +649,90 @@ goodix_rx_cb (FpiUsbTransfer *transfer,
 
       /* Native EC has no response event. Optional category-A/command-7 data
        * is ignored independently of the command or event currently awaited. */
-      if (goodix_proto_rx_parse (&self->rx, &category, &command,
-                                 &payload, &payload_len) &&
+      if (self->reply_valid &&
           category == 0x0a && command == 7)
         {
           goodix_proto_rx_reset (&self->rx);
           goto receive_more;
         }
 
-      if (self->cmd_ssm == transfer->ssm)
-        {
-          GoodixCmdOperation *cmd_operation = fpi_ssm_get_data (transfer->ssm);
-
-          if (cmd_operation->phase == GOODIX_CMD_RECV_ACK &&
-              goodix_proto_rx_parse (&self->rx, &category, &command,
-                                     &payload, &payload_len) &&
-              category == GOODIX_PROTO_CATEGORY_ACK && command == GOODIX_PROTO_CMD_ACK &&
-              payload_len >= 2)
-            expected_ack = payload[0] == GOODIX_PROTO_CMD_BYTE (
-              cmd_operation->cmd.category, cmd_operation->cmd.command);
-        }
       /* Native updates the acknowledged command's independent slot. A late
        * ACK from a repeated mode command cannot satisfy a different command
        * or an event/data wait. Validate the envelope before routing it. */
-      if (!expected_ack &&
-          goodix_proto_rx_parse (&self->rx, &category, &command,
-                                 &payload, &payload_len) &&
-          category == GOODIX_PROTO_CATEGORY_ACK && command == GOODIX_PROTO_CMD_ACK &&
-          payload_len >= 2 &&
+      if (ack_packet && !expected_ack &&
           (self->retried_mode_acks & goodix_mode_ack_bit (payload[0])))
         {
           goodix_proto_rx_reset (&self->rx);
           goto receive_more;
         }
-      /* Message complete — advance SSM */
-      goodix_recv_operation_finish (dev, transfer->ssm, operation);
-      goodix_recv_operation_free (operation);
-      fpi_ssm_next_state (transfer->ssm);
+      if (ack_wait)
+        {
+          guint8 status;
+
+          if (!goodix_validate_ack_for_cmd (dev, current, &status, &error))
+            goodix_transport_complete (operation, error);
+          /* Even status (or drained cleanup FDT) preserves this attempt's
+           * original ACK deadline. Only exhaustion permits a retry. */
+          else if (!(status & GOODIX_PROTO_ACK_FLAG_VALID))
+            {
+              goodix_proto_rx_reset (&self->rx);
+              goto receive_more;
+            }
+          else
+            {
+              current->ack_status = status;
+              if (!current->expect_data ||
+                  (current->response_bit & self->command_response_ready))
+                goodix_transport_complete (operation, NULL);
+              else
+                {
+                  operation->phase = GOODIX_TRANSPORT_RESPONSE;
+                  operation->deadline_us = g_get_monotonic_time () +
+                                           operation->response_timeout_ms * 1000LL;
+                  g_clear_object (&operation->cancellable);
+                  if (operation->response_bit == 4)
+                    operation->cancellable = g_object_ref (fpi_device_get_cancellable (dev));
+                  goodix_transport_receive (operation, operation->response_timeout_ms);
+                }
+            }
+          return;
+        }
+      if (operation->phase == GOODIX_TRANSPORT_EVENT)
+        {
+          GoodixFdtNotification *pending = &self->pending_fdt;
+          if (!goodix_cmd_parse_fdt_event (dev, operation->event_mode,
+                                          &pending->type, &pending->event, &error))
+            goto protocol_error;
+          goodix_recv_apply_fdt_event (dev, pending->type, &pending->event);
+          memcpy (pending->prior_down, self->fdt_prior_down, sizeof (pending->prior_down));
+        }
+      goodix_transport_complete (operation, NULL);
     }
   else
     {
       /* Preserve unrelated per-continuation data budgets. Scoped command waits
        * keep one deadline across interleaved packets and continuations. */
-      if (!operation->fixed_deadline && self->idle_rx_ssm != transfer->ssm)
+      if (!fixed_deadline && !goodix_transport_is_idle (operation))
         operation->deadline_us = g_get_monotonic_time () + GOODIX_DATA_TIMEOUT * 1000LL;
       goto receive_more;
     }
   return;
 
+protocol_error:
+  goodix_transport_complete (operation, error);
+  return;
+
 receive_more:
   {
     if (self->rx.len == 0)
-      self->rx_idle_partial = FALSE;
-    if (self->idle_rx_ssm == transfer->ssm &&
+      {
+        self->rx_idle_partial = FALSE;
+        self->reply_valid = FALSE;
+      }
+    if (goodix_transport_is_idle (operation) &&
         g_cancellable_is_cancelled (operation->cancellable))
       {
-        goodix_recv_operation_finish (dev, transfer->ssm, operation);
-        goodix_recv_operation_free (operation);
-        fpi_ssm_mark_completed (transfer->ssm);
+        goodix_transport_complete (operation, NULL);
         return;
       }
     guint timeout = operation->deadline_us ?
@@ -879,36 +746,46 @@ receive_more:
         return;
       }
     next = fpi_usb_transfer_new (dev);
-    next->ssm = transfer->ssm;
     fpi_usb_transfer_fill_bulk (next, GOODIX_EP_IN, GOODIX_USB_CHUNK_SIZE);
     fpi_usb_transfer_submit (next, timeout, operation->cancellable,
                              goodix_rx_cb, operation);
   }
 }
 
-gboolean
-goodix_recv_start_cancellable_full (
-  FpiSsm                     *ssm,
-  FpDevice                   *dev,
-  GCancellable               *cancellable,
-  GoodixRecvCancelledCallback cancelled_cb,
-  gpointer                    user_data)
+void
+goodix_transport_invalidate (FpDevice *dev)
 {
-  return goodix_recv_start_full (ssm, dev, 0, cancellable,
-                                 cancelled_cb, user_data);
+  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
+
+  g_assert (!self->transport);
+  if (self->rx.buf)
+    goodix_proto_rx_reset (&self->rx);
+  self->reply_valid = FALSE;
+  self->rx_idle_partial = FALSE;
+  self->pending_fdt.event.pending = FALSE;
 }
 
 gboolean
-goodix_recv_take_pending_fdt (FpDevice *dev)
+goodix_recv_select_fdt (FpDevice *dev, GoodixFdtEventType *type,
+                        GoodixProfile9FdtEvent *event,
+                        guint16 *prior_down, GError **error)
 {
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  gsize len = self->pending_fdt_packet_len;
+  GoodixFdtNotification *pending = &self->pending_fdt;
 
-  if (!len || self->rx_active || self->cmd_owner)
-    return FALSE;
-  self->pending_fdt_packet_len = 0;
-  goodix_proto_rx_reset (&self->rx);
-  return goodix_proto_rx_feed_chunk (&self->rx, self->pending_fdt_packet, len);
+  g_assert (!self->transport);
+  if (pending->event.pending)
+    {
+      *event = pending->event;
+      *type = pending->type;
+      if (*type == GOODIX_FDT_EVENT_REVERSE)
+        memcpy (prior_down, pending->prior_down, sizeof (pending->prior_down));
+      pending->event.pending = FALSE;
+      return TRUE;
+    }
+  g_set_error_literal (error, FP_DEVICE_ERROR, FP_DEVICE_ERROR_PROTO,
+                       "No completed FDT event");
+  return FALSE;
 }
 
 void
@@ -937,265 +814,131 @@ goodix_recv_apply_fdt_event (FpDevice                     *dev,
     }
 }
 
-/* ========================================================================
- * Command sub-SSM: send → recv ACK → recv data
- * ======================================================================== */
-
-static void
-goodix_cmd_ssm_handler (FpiSsm   *ssm,
-                        FpDevice *dev)
-{
-  GoodixCmdOperation *operation = fpi_ssm_get_data (ssm);
-  GoodixCmd *cmd = &operation->cmd;
-
-  operation->phase = fpi_ssm_get_cur_state (ssm);
-
-  switch (fpi_ssm_get_cur_state (ssm))
-    {
-    case GOODIX_CMD_SEND:
-      operation->attempt++;
-      FPI_DEVICE_GOODIX53X5 (dev)->command_response_ready &= ~operation->response_bit;
-      if (operation->response_bit)
-        FPI_DEVICE_GOODIX53X5 (dev)->retried_mode_acks |= goodix_mode_ack_bit (
-          GOODIX_PROTO_CMD_BYTE (cmd->category, cmd->command));
-      goodix_send_message (ssm, dev, cmd->category, cmd->command,
-                           cmd->payload, cmd->payload_len, cmd->use_checksum);
-      break;
-
-    case GOODIX_CMD_RECV_ACK:
-      {
-        guint timeout;
-
-        if (!operation->ack_deadline_us)
-          operation->ack_deadline_us = g_get_monotonic_time () + operation->ack_timeout_ms * 1000LL;
-        timeout = goodix_deadline_remaining (operation->ack_deadline_us);
-        if (!timeout)
-          {
-            GError *error = g_error_new_literal (
-              G_USB_DEVICE_ERROR, G_USB_DEVICE_ERROR_TIMED_OUT, "ACK deadline expired");
-
-            if (!goodix_cmd_retry (dev, ssm, error))
-              fpi_ssm_mark_failed (ssm, error);
-            return;
-          }
-        goodix_recv_start (ssm, dev, timeout,
-                          ((cmd->category == 0 && cmd->command == 0) ||
-                           (cmd->category == 0x0a && cmd->command == 4)) ?
-                          fpi_device_get_cancellable (dev) : NULL);
-      }
-      break;
-
-    case GOODIX_CMD_VALIDATE_ACK:
-      {
-        g_autoptr(GError) error = NULL;
-        guint8 status;
-
-        if (operation->cancelled_fdt_mode != GOODIX_PROFILE9_FDT_WAIT_NONE &&
-            goodix_try_drain_cancelled_fdt (dev, operation, &error))
-          {
-            fpi_ssm_jump_to_state (ssm, GOODIX_CMD_RECV_ACK);
-            return;
-          }
-        if (error)
-          {
-            fpi_ssm_mark_failed (ssm, g_steal_pointer (&error));
-            return;
-          }
-
-        if (!goodix_validate_ack_for_cmd (dev, cmd, &status, &error))
-          {
-            fpi_ssm_mark_failed (ssm, g_steal_pointer (&error));
-            return;
-          }
-
-        /* A valid even status leaves the ACK unsatisfied. Keep this attempt's
-         * original deadline; only exhaustion invokes its existing retry policy. */
-        if ((status & GOODIX_PROTO_ACK_FLAG_VALID) == 0)
-          {
-            fpi_ssm_jump_to_state (ssm, GOODIX_CMD_RECV_ACK);
-          }
-        else
-          {
-            operation->ack_status = status;
-            fpi_ssm_next_state (ssm);
-          }
-      }
-      break;
-
-    case GOODIX_CMD_RECV_DATA:
-      if (operation->response_bit & FPI_DEVICE_GOODIX53X5 (dev)->command_response_ready)
-        fpi_ssm_next_state (ssm);
-      else
-        goodix_recv_start (ssm, dev, operation->response_timeout_ms,
-                          operation->response_bit == 4 ? fpi_device_get_cancellable (dev) : NULL);
-      break;
-    }
-}
-
-static void
-goodix_cmd_ssm_done (FpiSsm   *ssm,
-                     FpDevice *dev,
-                     GError   *error)
+void
+goodix_transport_command (FpDevice                     *dev,
+                           const GoodixTransportRequest *request,
+                           GoodixTransportDone          done,
+                           gpointer                     data)
 {
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  GoodixCmdOperation *operation = fpi_ssm_get_data (ssm);
+  GoodixTransport *operation = self->transport;
+  guint8 category = request->cmd.category, command = request->cmd.command;
 
-  g_assert (self->cmd_ssm == ssm);
-  self->cmd_owner = NULL;
-  self->cmd_ssm = NULL;
-
-  if (error)
+  if (operation && operation->phase != GOODIX_TRANSPORT_IDLE)
     {
-      g_prefix_error (&error,
-                      "Command cat=0x%02x cmd=0x%02x phase=%s attempt=%u parent-state=%d: ",
-                      operation->cmd.category, operation->cmd.command,
-                      goodix_cmd_phase_name (operation->phase), operation->attempt,
-                      fpi_ssm_get_cur_state (operation->parent_ssm));
-    }
-  if (operation->result_cb)
-    {
-      operation->result_cb (operation->parent_ssm, dev, operation->ack_status,
-                            error && goodix_cmd_native_zero (operation, error), error);
+      GoodixTransportResult result = {0};
+      done (dev, &result, fpi_device_error_new_msg (FP_DEVICE_ERROR_BUSY,
+             "A device transport operation is already active"), data);
       return;
     }
 
-  if (error)
-    {
-      goodix_scan_note_command_error (operation->parent_ssm, dev, error);
-      goodix_mark_coordinator_io_failure (self, error);
-      fpi_ssm_mark_failed (operation->parent_ssm, error);
-    }
-  else
-    {
-      if (operation->idle_after_ack)
-        goodix_idle_recv_start (dev);
-      fpi_ssm_next_state (operation->parent_ssm);
-    }
-}
-
-static void
-goodix_cmd_begin_after_idle (FpDevice *dev, gpointer data)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-
-  fpi_ssm_start (self->cmd_ssm, goodix_cmd_ssm_done);
-}
-
-static void
-goodix_run_cmd_full (FpiSsm                    *parent_ssm,
-                     FpDevice                  *dev,
-                     guint8                     category,
-                     guint8                     command,
-                     const guint8              *payload,
-                     gsize                      payload_len,
-                     gboolean                   expect_data,
-                     GoodixProfile9FdtWaitMode  cancelled_mode,
-                     GoodixCmdResultCallback   callback,
-                     gboolean                  idle_after_ack)
-{
-  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
-  FpiSsm *cmd_ssm;
-  GoodixCmdOperation *operation;
-  GoodixCmd *cmd;
-
-  if (self->cmd_owner || (self->rx_active && self->rx_owner != self->idle_rx_ssm))
-    {
-      fpi_ssm_mark_failed (
-        parent_ssm,
-        fpi_device_error_new_msg (
-          FP_DEVICE_ERROR_BUSY,
-          self->cmd_owner ? "A device command is already active"
-                          : "Cannot start a command while a receive is active"));
-      return;
-    }
-
-  operation = g_new0 (GoodixCmdOperation, 1);
-  operation->parent_ssm = parent_ssm;
-  operation->result_cb = callback;
-  operation->idle_after_ack = idle_after_ack;
-  operation->cancelled_fdt_mode = cancelled_mode;
-  operation->response_bit = expect_data ?
+  if (!operation)
+    operation = goodix_transport_new (dev, GOODIX_TRANSPORT_SEND, done, data);
+  operation->done = done;
+  operation->data = data;
+  operation->cmd = request->cmd;
+  operation->cmd.payload = request->cmd.payload_len ?
+    g_memdup2 (request->cmd.payload, request->cmd.payload_len) : NULL;
+  operation->expect_data = request->expect_data;
+  operation->idle_after_ack = request->idle_after_ack;
+  operation->cancelled_fdt_mode = request->cancelled_mode;
+  operation->response_bit = request->expect_data ?
                             (category == 3 && command == 3 ? 1 :
                              category == 9 && command == 0 ? 2 :
                              category == 0x0a && command == 4 ? 4 : 0) : 0;
-  operation->retry_mode = operation->response_bit || (!expect_data &&
+  operation->retry_mode = operation->response_bit || (!request->expect_data &&
                          ((category == GOODIX_PROTO_CATEGORY_FDT &&
                            (command == GOODIX_PROTO_CMD_FDT_DOWN || command == GOODIX_PROTO_CMD_FDT_UP)) ||
                           ((category == 0x06 || category == 0) && command == 0)));
-  cmd = &operation->cmd;
-  cmd->category = category;
-  cmd->command = command;
-  cmd->use_checksum = TRUE;
   goodix_cmd_set_budgets (operation);
 
-  if (payload_len > 0 && payload != NULL)
+  if (operation->phase == GOODIX_TRANSPORT_IDLE)
     {
-      cmd->payload = g_memdup2 (payload, payload_len);
-      cmd->payload_len = payload_len;
+      operation->phase = GOODIX_TRANSPORT_JOIN_IDLE;
+      g_cancellable_cancel (operation->cancellable);
     }
   else
+    goodix_transport_send (operation);
+}
+
+static void
+goodix_transport_wait (FpDevice                  *dev,
+                        GoodixTransportPhase       phase,
+                        guint                      timeout,
+                        GoodixProfile9FdtWaitMode  mode,
+                        GoodixTransportDone        done,
+                        gpointer                   data)
+{
+  FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
+  GoodixTransport *operation;
+
+  if (self->transport)
     {
-      cmd->payload = NULL;
-      cmd->payload_len = 0;
+      GoodixTransportResult result = {0};
+      done (dev, &result, fpi_device_error_new_msg (FP_DEVICE_ERROR_BUSY,
+             "A device transport operation is already active"), data);
+      return;
     }
-
-  self->cmd_owner = parent_ssm;
-
-  cmd_ssm = fpi_ssm_new_full (dev, goodix_cmd_ssm_handler,
-                               expect_data ? GOODIX_CMD_NUM_STATES : GOODIX_CMD_RECV_DATA,
-                                expect_data ? GOODIX_CMD_NUM_STATES : GOODIX_CMD_RECV_DATA,
-                                "goodix-cmd");
-  fpi_ssm_set_data (cmd_ssm, operation,
-                    (GDestroyNotify) goodix_cmd_operation_free);
-  self->cmd_ssm = cmd_ssm;
-  goodix_idle_recv_stop (dev, goodix_cmd_begin_after_idle, NULL);
+  operation = goodix_transport_new (dev, phase, done, data);
+  operation->event_mode = mode;
+  operation->deadline_us = timeout ? g_get_monotonic_time () + timeout * 1000LL : 0;
+  if (phase == GOODIX_TRANSPORT_EVENT)
+    {
+      operation->cancellable = g_cancellable_new ();
+      if (self->pending_fdt.event.pending)
+        {
+          goodix_transport_complete (operation, NULL);
+          return;
+        }
+    }
+  goodix_transport_receive (operation, timeout);
 }
 
 void
-goodix_run_cmd (FpiSsm       *parent_ssm,
-                FpDevice     *dev,
-                guint8        category,
-                guint8        command,
-                const guint8 *payload,
-                gsize         payload_len,
-                gboolean      expect_data)
+goodix_transport_wait_event (FpDevice                  *dev,
+                              GoodixProfile9FdtWaitMode mode,
+                              GoodixTransportDone       done,
+                              gpointer                  data)
 {
-  goodix_run_cmd_full (parent_ssm, dev, category, command, payload,
-                       payload_len, expect_data,
-                       GOODIX_PROFILE9_FDT_WAIT_NONE, NULL, FALSE);
+  goodix_transport_wait (dev, GOODIX_TRANSPORT_EVENT, 0, mode, done, data);
 }
 
 void
-goodix_run_cmd_ec_off (FpiSsm *ssm, FpDevice *dev,
-                       const guint8 *payload, gsize payload_len)
+goodix_transport_wait_reply (FpDevice           *dev,
+                              guint               timeout,
+                              GoodixTransportDone done,
+                              gpointer            data)
 {
-  goodix_run_cmd_full (ssm, dev, 0x0a, 7, payload, payload_len, FALSE,
-                       GOODIX_PROFILE9_FDT_WAIT_NONE, NULL, TRUE);
+  goodix_transport_wait (dev, GOODIX_TRANSPORT_REPLY, timeout,
+                         GOODIX_PROFILE9_FDT_WAIT_NONE, done, data);
 }
 
 void
-goodix_run_cmd_result (FpiSsm *ssm, FpDevice *dev,
-                       guint8 category, guint8 command,
-                       const guint8 *payload, gsize payload_len,
-                       gboolean expect_data, GoodixCmdResultCallback callback)
+goodix_transport_cancel_event (FpDevice *dev)
 {
-  goodix_run_cmd_full (ssm, dev, category, command, payload, payload_len,
-                       expect_data, GOODIX_PROFILE9_FDT_WAIT_NONE, callback, FALSE);
+  GoodixTransport *operation = FPI_DEVICE_GOODIX53X5 (dev)->transport;
+
+  if (operation && operation->phase == GOODIX_TRANSPORT_EVENT)
+    g_cancellable_cancel (operation->cancellable);
 }
 
 void
-goodix_run_cmd_drain_fdt_once (
-  FpiSsm                    *parent_ssm,
-  FpDevice                  *dev,
-  guint8                     category,
-  guint8                     command,
-  const guint8              *payload,
-  gsize                      payload_len,
-  GoodixProfile9FdtWaitMode  cancelled_mode)
+goodix_transport_quiesce (FpDevice             *dev,
+                           GoodixTransportJoined joined,
+                           gpointer              data)
 {
-  g_return_if_fail (cancelled_mode != GOODIX_PROFILE9_FDT_WAIT_NONE);
-  goodix_run_cmd_full (parent_ssm, dev, category, command, payload,
-                       payload_len, FALSE, cancelled_mode, NULL, FALSE);
+  GoodixTransport *operation = FPI_DEVICE_GOODIX53X5 (dev)->transport;
+
+  if (!operation)
+    {
+      joined (dev, data);
+      return;
+    }
+  g_assert (operation->phase == GOODIX_TRANSPORT_IDLE);
+  operation->phase = GOODIX_TRANSPORT_STOPPING;
+  operation->joined = joined;
+  operation->joined_data = data;
+  g_cancellable_cancel (operation->cancellable);
 }
 
 gboolean
@@ -1208,9 +951,14 @@ goodix_parse_reply (FpDevice      *dev,
 {
   FpiDeviceGoodix53x5 *self = FPI_DEVICE_GOODIX53X5 (dev);
 
-  if (goodix_proto_rx_parse (&self->rx, out_category, out_command,
-                             out_payload, out_payload_len))
-    return TRUE;
+  if (self->reply_valid)
+    {
+      *out_category = self->reply_category;
+      *out_command = self->reply_command;
+      *out_payload = self->reply_payload;
+      *out_payload_len = self->reply_payload_len;
+      return TRUE;
+    }
 
   g_set_error_literal (error, FP_DEVICE_ERROR, FP_DEVICE_ERROR_PROTO,
                        "Failed to parse device reply");
